Refactor Spotify API client into monadic structures
The goal of this refactoring was to encapsulate the results of calls to
the Spotify API into a monadic `Results` structure, so that calls could
be chained using `>>=`. The signatures of various API calls would also
require client code to authorize and call endpoints in the proper order.
These goals were not fully realized, as calls to the API endpoints are
still wrapped in the IO monad, as well as the Result monad, but even so,
this refactoring cleans up the client API, making it easier to work
with.